add JSON post request features
The post request code in Plotter.js was updated to successfully post a backend/sampleData/example_fasta.json-formatted JSON file to the backend. A rough draft of the composite data version of the button is also included in this commit. Two text box areas display the JSON objects that were selected to upload.

These changes will still need to be cleaned up for efficiency, error-checking, generalizability, and there are some minor frontend update issues that may need to be addressed.