fix JSON uploads in Plotter.js
Loop through Objects in uploaded JSON file to post to backend was adjusted to loop on keys rather than position index.

This fix is needed because the prior assumption was the input JSON was an array of FASTA or composite data objects rather than an object with each new to-post data object as the value for a unique key.

Before assumed: [ {FASTA_1_info}, {FASTA_2_info},...]

Now assumed: { "0":{FASTA_1_info}, "1":{FASTA_2_info},...}