Add incomplete upload composite file parser
This commit includes updates to the Plotter.js Component so that the upload button can read the file and parse the tab-delimited composite results (formatted to ScriptManager's TagPileup output) into value arrays. Next step is to format these arrays into a JSON for posting to the database and then posting the data. There is some placeholder pseudocode indicating where in the files this should be written.

The helloworld.js placeholder has been removed and replaced with parseComposites.js. Plotter.js now calls the new parseComposites.js on each file returned from the input File Selector and prints the value arrays to the console.